+++ app.js
@@ -14,6 +14,7 @@ const presidentsRouter = require('./routes/presidents');
 const parastatalsRouter = require('./routes/parastatals');
 const lakesRouter = require('./routes/lakes');
 const riversRouter = require('./routes/rivers');
+const healthFacilitiesRouter = require('./routes/health-facilities');
 const searchRouter = require('./routes/search');
 const notFound = require('./middleware/notFound');
 const errorHandler = require('./middleware/errorHandler');
@@ -37,6 +38,7 @@ app.use('/api/v1/presidents', presidentsRouter);
 app.use('/api/v1/parastatals', parastatalsRouter);
 app.use('/api/v1/lakes', lakesRouter);
 app.use('/api/v1/rivers', riversRouter);
+app.use('/api/v1/health-facilities', healthFacilitiesRouter);
 app.use('/api/v1/search', searchRouter);
 
 app.get('/', (req, res) => {